Update dependencies in HF Text inference and imports
