+++ llama-index-integrations/embeddings/llama-index-embeddings-huggingface/llama_index/embeddings/huggingface/base.py
@@ -17,7 +17,7 @@ from llama_index.core.bridge.pydantic import Field, PrivateAttr
 from llama_index.core.callbacks import CallbackManager
 from llama_index.core.utils import get_cache_dir, infer_torch_device
 from llama_index.embeddings.huggingface.pooling import Pooling
-from llama_index.embeddings.huggingface.utils import (
+from llama_index.core.embeddings.utils import (
     DEFAULT_HUGGINGFACE_EMBEDDING_MODEL,
     format_query,
     format_text,
