+++ llama-index-integrations/embeddings/llama-index-embeddings-text-embeddings-inference/pyproject.toml
@@ -32,7 +32,6 @@ version = "0.1.2"
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
 llama-index-core = "^0.10.1"
-llama-index-embeddings-huggingface = "^0.1.1"
 httpx = ">=0.26.0"
 
 [tool.poetry.group.dev.dependencies]
